UI: color/B&W toggle, optional PDF title, full weekday headers, dup-name flag
Round of display refinements on top of Phase 1:

- Color / black-and-white toggle (persisted). B&W renders every cell white with
  black text so prints are clean and shift times stay legible; color keeps the
  time-of-day fills.
- Optional schedule title: an editable heading that prints as the document
  title and drives document.title (the PDF's default filename); blank falls
  back to "Shift Schedule".
- Day headers now show the full weekday name (Monday) over the date.
- Duplicate person names are flagged with a non-blocking, no-print marker.
- Cell dropdown lists start times first, then Off/PTO/Holiday.
- Week title date range is now bold/dark for legibility.
- Heavier cell text for readability on the color fills.

Model: Schedule gains an optional `title`; SET_TITLE action; storage guard
validates it. 77 tests / typecheck / lint / build all green.

+++ src/lib/dates.ts
@@ -70,9 +70,9 @@ export function mondayOf(iso: string): string {
 }
 
 /**
- * Split a date into the two lines a column header shows: a short weekday
- * (`Mon`) over a short month/day (`Jul 27`). Fixed to en-US so the header reads
- * the same everywhere — this is a document, not a localized UI.
+ * Split a date into the two lines a column header shows: the full weekday
+ * (`Monday`) over a short month/day (`Jul 27`). Fixed to en-US so the header
+ * reads the same everywhere — this is a document, not a localized UI.
  */
 export function formatDayHeader(iso: string): {
   weekday: string;
@@ -80,7 +80,7 @@ export function formatDayHeader(iso: string): {
 } {
   const date = parseISODateLocal(iso);
   return {
-    weekday: date.toLocaleDateString('en-US', { weekday: 'short' }),
+    weekday: date.toLocaleDateString('en-US', { weekday: 'long' }),
     date: date.toLocaleDateString('en-US', { month: 'short', day: 'numeric' }),
   };
 }
